backend/requirements.txt
fastapi
uvicorn
ultralytics
pillow
python-multipart
httpx
